requirements.txt
asgiref==3.7.2
boto3==1.34.27
django-cors-headers==4.3.1
django-filter==23.5
Django==5.0.1
djangorestframework-simplejwt==5.3.1
djangorestframework==3.14.0
mysqlclient==2.2.1
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
